move mock classes into mock package
