+++ dubbo-config/dubbo-config-api/src/test/java/com/alibaba/dubbo/config/mock/MockProtocol2.java
@@ -15,16 +15,34 @@
  * limitations under the License.
  */
 
-package com.alibaba.dubbo.config;
+package com.alibaba.dubbo.config.mock;
 
 import com.alibaba.dubbo.common.URL;
-import com.alibaba.dubbo.common.threadpool.ThreadPool;
+import com.alibaba.dubbo.rpc.Exporter;
+import com.alibaba.dubbo.rpc.Invoker;
+import com.alibaba.dubbo.rpc.Protocol;
+import com.alibaba.dubbo.rpc.RpcException;
 
-import java.util.concurrent.Executor;
+public class MockProtocol2 implements Protocol {
+    public static Protocol delegate;
 
-public class MockThreadPool implements ThreadPool  {
     @Override
-    public Executor getExecutor(URL url) {
-        return null;
+    public int getDefaultPort() {
+        return delegate.getDefaultPort();
+    }
+
+    @Override
+    public <T> Exporter<T> export(Invoker<T> invoker) throws RpcException {
+        return delegate.export(invoker);
+    }
+
+    @Override
+    public <T> Invoker<T> refer(Class<T> type, URL url) throws RpcException {
+        return delegate.refer(type, url);
+    }
+
+    @Override
+    public void destroy() {
+        delegate.destroy();
     }
 }
