+++ dubbo-config/dubbo-config-api/src/test/java/com/alibaba/dubbo/config/mock/MockTelnetHandler.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package com.alibaba.dubbo.config;
+package com.alibaba.dubbo.config.mock;
 
 import com.alibaba.dubbo.remoting.Channel;
 import com.alibaba.dubbo.remoting.RemotingException;
