+++ dubbo-config/dubbo-config-api/src/test/resources/META-INF/services/com.alibaba.dubbo.common.status.StatusChecker
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-mockstatuschecker=com.alibaba.dubbo.config.MockStatusChecker
+mockstatuschecker=com.alibaba.dubbo.config.mock.MockStatusChecker
