+++ dubbo-config/dubbo-config-api/src/test/resources/META-INF/services/com.alibaba.dubbo.registry.RegistryFactory
@@ -1,2 +1,2 @@
-mockregistry=com.alibaba.dubbo.config.support.MockRegistryFactory
-mockprotocol2=com.alibaba.dubbo.config.MockRegistryFactory
+mockregistry=com.alibaba.dubbo.config.mock.MockRegistryFactory
+mockprotocol2=com.alibaba.dubbo.config.mock.MockRegistryFactory2
