+++ dubbo-config/dubbo-config-api/src/test/resources/META-INF/services/com.alibaba.dubbo.remoting.Codec
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-mockcodec=com.alibaba.dubbo.config.MockCodec
+mockcodec=com.alibaba.dubbo.config.mock.MockCodec
