+++ dubbo-config/dubbo-config-api/src/test/resources/META-INF/services/com.alibaba.dubbo.remoting.Transporter
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-mocktransporter=com.alibaba.dubbo.config.MockTransporter
+mocktransporter=com.alibaba.dubbo.config.mock.MockTransporter
