+++ dubbo-config/dubbo-config-api/src/test/resources/META-INF/services/com.alibaba.dubbo.remoting.exchange.Exchanger
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-mockexchanger=com.alibaba.dubbo.config.MockExchanger
+mockexchanger=com.alibaba.dubbo.config.mock.MockExchanger
