+++ dubbo-config/dubbo-config-api/src/test/resources/META-INF/services/com.alibaba.dubbo.remoting.telnet.TelnetHandler
@@ -14,5 +14,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-mocktelnethandler=com.alibaba.dubbo.config.MockTelnetHandler
+mocktelnethandler=com.alibaba.dubbo.config.mock.MockTelnetHandler
