+++ dubbo-config/dubbo-config-api/src/test/resources/META-INF/services/com.alibaba.dubbo.rpc.Filter
@@ -1,1 +1,1 @@
-mockfilter=com.alibaba.dubbo.config.MockFilter
+mockfilter=com.alibaba.dubbo.config.mock.MockFilter
