+++ dubbo-config/dubbo-config-api/src/test/resources/META-INF/services/com.alibaba.dubbo.rpc.InvokerListener
@@ -1,1 +1,1 @@
-mockinvokerlistener=com.alibaba.dubbo.config.MockInvokerListener
+mockinvokerlistener=com.alibaba.dubbo.config.mock.MockInvokerListener
