+++ dubbo-config/dubbo-config-api/src/test/resources/META-INF/services/com.alibaba.dubbo.rpc.ProxyFactory
@@ -1,1 +1,1 @@
-mockproxyfactory=com.alibaba.dubbo.config.MockProxyFactory
+mockproxyfactory=com.alibaba.dubbo.config.mock.MockProxyFactory
