+++ dubbo-config/dubbo-config-api/src/test/resources/META-INF/services/com.alibaba.dubbo.rpc.cluster.Cluster
@@ -1,1 +1,1 @@
-mockcluster=com.alibaba.dubbo.config.MockCluster
+mockcluster=com.alibaba.dubbo.config.mock.MockCluster
