+++ dubbo-config/dubbo-config-api/src/test/resources/META-INF/services/com.alibaba.dubbo.rpc.cluster.LoadBalance
@@ -1,1 +1,1 @@
-mockloadbalance=com.alibaba.dubbo.config.MockLoadBalance
+mockloadbalance=com.alibaba.dubbo.config.mock.MockLoadBalance
